+++ src/tests/specs/observatory/observatory.openid.upgrade.spec.ts
@@ -6,7 +6,7 @@ import { fromNullable } from '@dfinity/utils';
 import { inject } from 'vitest';
 import { mockCertificateDate, mockClientId } from '../../mocks/jwt.mocks';
 import { FETCH_CERTIFICATE_INTERVAL } from '../../mocks/observatory.mocks';
-import { makeMockGoogleOpenIdJwt } from '../../utils/jwt-test.utils';
+import { makeMockGoogleOpenIdJwt } from '../../utils/jwt-tests.utils';
 import { assertOpenIdHttpsOutcalls } from '../../utils/observatory-openid-tests.utils';
 import { tick } from '../../utils/pic-tests.utils';
 import { OBSERVATORY_WASM_PATH } from '../../utils/setup-tests.utils';
