+++ src/tests/utils/jwt-assemble-tests.utils.ts
@@ -1,5 +1,5 @@
 import { uint8ArrayToBase64 } from '@dfinity/utils';
-import type { MockOpenIdJwt } from './jwt-test.utils';
+import type { MockOpenIdJwt } from './jwt-tests.utils';
 
 export const toBase64URL = (uint8Array: Uint8Array): string =>
 	uint8ArrayToBase64(uint8Array).replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '');
